neighbour test update + resolverinterface added highest id

+++ src/NameServer/ResolverInterface.java
@@ -15,7 +15,7 @@ public interface ResolverInterface extends Remote
 	 * @return
 	 * @throws RemoteException
 	 */
-	public String getIP(int nodeId) throws RemoteException;
+	public String getIP(short nodeId) throws RemoteException;
 
 	/**
 	 * For a given filename it returns the IP address from the owner of the file
@@ -42,6 +42,6 @@ public interface ResolverInterface extends Remote
 
 	public void addToTree(short ID, String IP) throws RemoteException;   // todo: Remove when testing is complete
 
-	public void removeFromTree(int nodeId) throws RemoteException; //todo: Remove when testing complete
+	public void removeFromTree(short nodeId) throws RemoteException; //todo: Remove when testing complete
 
 }
